# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a assignment for DTS-Hacktiv8 final project.",
        "title": "Final Project API",
        "contact": {
            "name": "Aditya Gocendra",
            "email": "gocendra123@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/comments": {
            "post": {
                "security": [
                    {
                        "Bearer \u003cJWT\u003e": []
                    }
                ],
                "description": "Note : To create comment, you need to send the token that has been obtained from the login response with the format Bearer \u003cYOUR TOKEN\u003e.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comment"
                ],
                "summary": "Create Comment",
                "parameters": [
                    {
                        "description": "create comment",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CommentReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.CommentRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    }
                }
            }
        },
        "/comments/": {
            "get": {
                "security": [
                    {
                        "Bearer \u003cJWT\u003e": []
                    }
                ],
                "description": "Note : To get comment, you need to send the token that has been obtained from the login response with the format Bearer \u003cYOUR TOKEN\u003e.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comment"
                ],
                "summary": "Get Comments",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.GetCommentRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    }
                }
            }
        },
        "/comments/{commentid}": {
            "put": {
                "security": [
                    {
                        "Bearer \u003cJWT\u003e": []
                    }
                ],
                "description": "Note : To update comment, you need to send the token that has been obtained from the login response with the format Bearer \u003cYOUR TOKEN\u003e.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comment"
                ],
                "summary": "Update Comment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "commentid",
                        "name": "commentid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "udpate comment",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateCommentReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.UpdatePhotoRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer \u003cJWT\u003e": []
                    }
                ],
                "description": "Note : To delete comment, you need to send the token that has been obtained from the login response with the format Bearer \u003cYOUR TOKEN\u003e.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comment"
                ],
                "summary": "Delete Comment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "commentid",
                        "name": "commentid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.SuccessMessage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorMessage"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.CommentPhotoRes": {
            "type": "object",
            "properties": {
                "caption": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "photoUrl": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
        "dto.CommentReq": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "photo_id": {
                    "type": "integer"
                }
            }
        },
        "dto.CommentRes": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "photoID": {
                    "type": "integer"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
        "dto.CommentUserRes": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "dto.ErrorMessage": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "typeError": {
                    "type": "string"
                }
            }
        },
        "dto.GetCommentRes": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "photo": {
                    "$ref": "#/definitions/dto.CommentPhotoRes"
                },
                "photoID": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/dto.CommentUserRes"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
...
        "dto.SuccessMessage": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.UpdateCommentReq": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.UpdatePhotoRes": {
            "type": "object",
            "properties": {
                "caption": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "photoUrl": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer \u003cJWT\u003e": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}